refactor(monarchmoney): extract GraphQL queries and simplify params
- Replaced standard error prints with standard logging to prevent stderr pollution.
- Extracted GraphQL queries into a new `moneyflow/queries.py` file to separate data definitions from client logic.
- Simplified verbose variable mapping in `get_transactions` and `update_transaction` using dictionary comprehensions.
- Replaced manual date math with `dateutil.relativedelta`.
- Removed unnecessary `(object)` inheritance in classes.
