+++ moneyflow/queries.py
@@ -1,0 +1,1661 @@
+from gql import gql
+
+GET_ACCOUNTS = gql(
+    """
+          query GetAccounts {
+            accounts {
+              ...AccountFields
+              __typename
+            }
+            householdPreferences {
+              id
+              accountGroupOrder
+              __typename
+            }
+          }
+
+          fragment AccountFields on Account {
+            id
+            displayName
+            syncDisabled
+            deactivatedAt
+            isHidden
+            isAsset
+            mask
+            createdAt
+            updatedAt
+            displayLastUpdatedAt
+            currentBalance
+            displayBalance
+            includeInNetWorth
+            hideFromList
+            hideTransactionsFromReports
+            includeBalanceInNetWorth
+            includeInGoalBalance
+            dataProvider
+            dataProviderAccountId
+            isManual
+            transactionsCount
+            holdingsCount
+            manualInvestmentsTrackingMethod
+            order
+            logoUrl
+            type {
+              name
+              display
+              __typename
+            }
+            subtype {
+              name
+              display
+              __typename
+            }
+            credential {
+              id
+              updateRequired
+              disconnectedFromDataProviderAt
+              dataProvider
+              institution {
+                id
+                plaidInstitutionId
+                name
+                status
+                __typename
+              }
+              __typename
+            }
+            institution {
+              id
+              name
+              primaryColor
+              url
+              __typename
+            }
+            __typename
+          }
+        """
+)
+
+GET_ACCOUNT_TYPE_OPTIONS = gql(
+    """
+            query GetAccountTypeOptions {
+                accountTypeOptions {
+                    type {
+                        name
+                        display
+                        group
+                        possibleSubtypes {
+                            display
+                            name
+                            __typename
+                        }
+                        __typename
+                    }
+                    subtype {
+                        name
+                        display
+                        __typename
+                    }
+                    __typename
+                }
+            }
+        """
+)
+
+GET_ACCOUNT_RECENT_BALANCES = gql(
+    """
+            query GetAccountRecentBalances($startDate: Date!) {
+                accounts {
+                    id
+                    recentBalances(startDate: $startDate)
+                    __typename
+                }
+            }
+        """
+)
+
+GET_SNAPSHOTS_BY_ACCOUNT_TYPE = gql(
+    """
+            query GetSnapshotsByAccountType($startDate: Date!, $timeframe: Timeframe!) {
+                snapshotsByAccountType(startDate: $startDate, timeframe: $timeframe) {
+                    accountType
+                    month
+                    balance
+                    __typename
+                }
+                accountTypes {
+                    name
+                    group
+                    __typename
+                }
+            }
+        """
+)
+
+GET_AGGREGATE_SNAPSHOTS = gql(
+    """
+            query GetAggregateSnapshots($filters: AggregateSnapshotFilters) {
+                aggregateSnapshots(filters: $filters) {
+                    date
+                    balance
+                    __typename
+                }
+            }
+        """
+)
+
+WEB__CREATE_MANUAL_ACCOUNT = gql(
+    """
+            mutation Web_CreateManualAccount($input: CreateManualAccountMutationInput!) {
+                createManualAccount(input: $input) {
+                    account {
+                        id
+                        __typename
+                    }
+                    errors {
+                        ...PayloadErrorFields
+                        __typename
+                    }
+                __typename
+               }
+            }
+            fragment PayloadErrorFields on PayloadError {
+                fieldErrors {
+                    field
+                    messages
+                    __typename
+                }
+                message
+                code
+                __typename
+            }
+            """
+)
+
+COMMON__UPDATE_ACCOUNT = gql(
+    """
+            mutation Common_UpdateAccount($input: UpdateAccountMutationInput!) {
+                updateAccount(input: $input) {
+                    account {
+                        ...AccountFields
+                        __typename
+                    }
+                    errors {
+                        ...PayloadErrorFields
+                        __typename
+                    }
+                    __typename
+                }
+            }
+
+            fragment AccountFields on Account {
+                id
+                displayName
+                syncDisabled
+                deactivatedAt
+                isHidden
+                isAsset
+                mask
+                createdAt
+                updatedAt
+                displayLastUpdatedAt
+                currentBalance
+                displayBalance
+                includeInNetWorth
+                hideFromList
+                hideTransactionsFromReports
+                includeBalanceInNetWorth
+                includeInGoalBalance
+                dataProvider
+                dataProviderAccountId
+                isManual
+                transactionsCount
+                holdingsCount
+                manualInvestmentsTrackingMethod
+                order
+                icon
+                logoUrl
+                deactivatedAt
+                type {
+                    name
+                    display
+                    group
+                    __typename
+                }
+                subtype {
+                    name
+                    display
+                    __typename
+                }
+                credential {
+                    id
+                    updateRequired
+                    disconnectedFromDataProviderAt
+                    dataProvider
+                    institution {
+                        id
+                        plaidInstitutionId
+                        name
+                        status
+                        __typename
+                    }
+                    __typename
+                }
+                institution {
+                    id
+                    name
+                    primaryColor
+                    url
+                    __typename
+                }
+                __typename
+            }
+
+            fragment PayloadErrorFields on PayloadError {
+                fieldErrors {
+                    field
+                    messages
+                    __typename
+                }
+                message
+                code
+                __typename
+            }
+            """
+)
+
+COMMON__DELETE_ACCOUNT = gql(
+    """
+            mutation Common_DeleteAccount($id: UUID!) {
+                deleteAccount(id: $id) {
+                    deleted
+                    errors {
+                    ...PayloadErrorFields
+                    __typename
+                }
+                __typename
+                }
+            }
+            fragment PayloadErrorFields on PayloadError {
+                fieldErrors {
+                    field
+                    messages
+                    __typename
+                }
+                message
+                code
+                __typename
+            }
+            """
+)
+
+COMMON__FORCE_REFRESH_ACCOUNTS_MUTATION = gql(
+    """
+          mutation Common_ForceRefreshAccountsMutation($input: ForceRefreshAccountsInput!) {
+            forceRefreshAccounts(input: $input) {
+              success
+              errors {
+                ...PayloadErrorFields
+                __typename
+              }
+              __typename
+            }
+          }
+
+          fragment PayloadErrorFields on PayloadError {
+            fieldErrors {
+              field
+              messages
+              __typename
+            }
+            message
+            code
+            __typename
+          }
+          """
+)
+
+FORCE_REFRESH_ACCOUNTS_QUERY = gql(
+    """
+          query ForceRefreshAccountsQuery {
+            accounts {
+              id
+              hasSyncInProgress
+              __typename
+            }
+          }
+          """
+)
+
+WEB__GET_HOLDINGS = gql(
+    """
+          query Web_GetHoldings($input: PortfolioInput) {
+            portfolio(input: $input) {
+              aggregateHoldings {
+                edges {
+                  node {
+                    id
+                    quantity
+                    basis
+                    totalValue
+                    securityPriceChangeDollars
+                    securityPriceChangePercent
+                    lastSyncedAt
+                    holdings {
+                      id
+                      type
+                      typeDisplay
+                      name
+                      ticker
+                      closingPrice
+                      isManual
+                      closingPriceUpdatedAt
+                      __typename
+                    }
+                    security {
+                      id
+                      name
+                      type
+                      ticker
+                      typeDisplay
+                      currentPrice
+                      currentPriceUpdatedAt
+                      closingPrice
+                      closingPriceUpdatedAt
+                      oneDayChangePercent
+                      oneDayChangeDollars
+                      __typename
+                    }
+                    __typename
+                  }
+                  __typename
+                }
+                __typename
+              }
+              __typename
+            }
+          }
+        """
+)
+
+ACCOUNT_DETAILS_GET_ACCOUNT = gql(
+    """
+            query AccountDetails_getAccount($id: UUID!, $filters: TransactionFilterInput) {
+              account(id: $id) {
+                id
+                ...AccountFields
+                ...EditAccountFormFields
+                isLiability
+                credential {
+                  id
+                  hasSyncInProgress
+                  canBeForceRefreshed
+                  disconnectedFromDataProviderAt
+                  dataProvider
+                  institution {
+                    id
+                    plaidInstitutionId
+                    url
+                    ...InstitutionStatusFields
+                    __typename
+                  }
+                  __typename
+                }
+                institution {
+                  id
+                  plaidInstitutionId
+                  url
+                  ...InstitutionStatusFields
+                  __typename
+                }
+                __typename
+              }
+              transactions: allTransactions(filters: $filters) {
+                totalCount
+                results(limit: 20) {
+                  id
+                  ...TransactionsListFields
+                  __typename
+                }
+                __typename
+              }
+              snapshots: snapshotsForAccount(accountId: $id) {
+                date
+                signedBalance
+                __typename
+              }
+            }
+
+            fragment AccountFields on Account {
+              id
+              displayName
+              syncDisabled
+              deactivatedAt
+              isHidden
+              isAsset
+              mask
+              createdAt
+              updatedAt
+              displayLastUpdatedAt
+              currentBalance
+              displayBalance
+              includeInNetWorth
+              hideFromList
+              hideTransactionsFromReports
+              includeBalanceInNetWorth
+              includeInGoalBalance
+              dataProvider
+              dataProviderAccountId
+              isManual
+              transactionsCount
+              holdingsCount
+              manualInvestmentsTrackingMethod
+              order
+              logoUrl
+              type {
+                name
+                display
+                group
+                __typename
+              }
+              subtype {
+                name
+                display
+                __typename
+              }
+              credential {
+                id
+                updateRequired
+                disconnectedFromDataProviderAt
+                dataProvider
+                institution {
+                  id
+                  plaidInstitutionId
+                  name
+                  status
+                  __typename
+                }
+                __typename
+              }
+              institution {
+                id
+                name
+                primaryColor
+                url
+                __typename
+              }
+              __typename
+            }
+
+            fragment EditAccountFormFields on Account {
+              id
+              displayName
+              deactivatedAt
+              displayBalance
+              includeInNetWorth
+              hideFromList
+              hideTransactionsFromReports
+              dataProvider
+              dataProviderAccountId
+              isManual
+              manualInvestmentsTrackingMethod
+              isAsset
+              invertSyncedBalance
+              canInvertBalance
+              type {
+                name
+                display
+                __typename
+              }
+              subtype {
+                name
+                display
+                __typename
+              }
+              __typename
+            }
+
+            fragment InstitutionStatusFields on Institution {
+              id
+              hasIssuesReported
+              hasIssuesReportedMessage
+              plaidStatus
+              status
+              balanceStatus
+              transactionsStatus
+              __typename
+            }
+
+            fragment TransactionsListFields on Transaction {
+              id
+              ...TransactionOverviewFields
+              __typename
+            }
+
+            fragment TransactionOverviewFields on Transaction {
+              id
+              amount
+              pending
+              date
+              hideFromReports
+              plaidName
+              notes
+              isRecurring
+              reviewStatus
+              needsReview
+              dataProviderDescription
+              attachments {
+                id
+                __typename
+              }
+              isSplitTransaction
+              category {
+                id
+                name
+                group {
+                  id
+                  type
+                  __typename
+                }
+                __typename
+              }
+              merchant {
+                name
+                id
+                transactionsCount
+                __typename
+              }
+              tags {
+                id
+                name
+                color
+                order
+                __typename
+              }
+              __typename
+            }
+            """
+)
+
+WEB__GET_INSTITUTION_SETTINGS = gql(
+    """
+            query Web_GetInstitutionSettings {
+              credentials {
+                id
+                ...CredentialSettingsCardFields
+                __typename
+              }
+              accounts(filters: {includeDeleted: true}) {
+                id
+                displayName
+                subtype {
+                  display
+                  __typename
+                }
+                mask
+                credential {
+                  id
+                  __typename
+                }
+                deletedAt
+                __typename
+              }
+              subscription {
+                isOnFreeTrial
+                hasPremiumEntitlement
+                __typename
+              }
+            }
+
+            fragment CredentialSettingsCardFields on Credential {
+              id
+              updateRequired
+              disconnectedFromDataProviderAt
+              ...InstitutionInfoFields
+              institution {
+                id
+                name
+                url
+                __typename
+              }
+              __typename
+            }
+
+            fragment InstitutionInfoFields on Credential {
+              id
+              displayLastUpdatedAt
+              dataProvider
+              updateRequired
+              disconnectedFromDataProviderAt
+              ...InstitutionLogoWithStatusFields
+              institution {
+                id
+                name
+                hasIssuesReported
+                hasIssuesReportedMessage
+                __typename
+              }
+              __typename
+            }
+
+            fragment InstitutionLogoWithStatusFields on Credential {
+              dataProvider
+              updateRequired
+              institution {
+                hasIssuesReported
+                status
+                balanceStatus
+                transactionsStatus
+                __typename
+              }
+              __typename
+            }
+        """
+)
+
+COMMON__GET_JOINT_PLANNING_DATA = gql(
+    """
+            query Common_GetJointPlanningData($startDate: Date!, $endDate: Date!) {
+              budgetSystem
+              budgetData(startMonth: $startDate, endMonth: $endDate) {
+                ...BudgetDataFields
+                __typename
+              }
+              categoryGroups {
+                ...BudgetCategoryGroupFields
+                __typename
+              }
+              goalsV2 {
+                ...BudgetDataGoalsV2Fields
+                __typename
+              }
+            }
+            
+            fragment BudgetDataMonthlyAmountsFields on BudgetMonthlyAmounts {
+              month
+              plannedCashFlowAmount
+              plannedSetAsideAmount
+              actualAmount
+              remainingAmount
+              previousMonthRolloverAmount
+              rolloverType
+              cumulativeActualAmount
+              rolloverTargetAmount
+              __typename
+            }
+            
+            fragment BudgetMonthlyAmountsByCategoryFields on BudgetCategoryMonthlyAmounts {
+              category {
+                id
+                __typename
+              }
+              monthlyAmounts {
+                ...BudgetDataMonthlyAmountsFields
+                __typename
+              }
+              __typename
+            }
+            
+            fragment BudgetMonthlyAmountsByCategoryGroupFields on BudgetCategoryGroupMonthlyAmounts {
+              categoryGroup {
+                id
+                __typename
+              }
+              monthlyAmounts {
+                ...BudgetDataMonthlyAmountsFields
+                __typename
+              }
+              __typename
+            }
+            
+            fragment BudgetMonthlyAmountsForFlexExpenseFields on BudgetFlexMonthlyAmounts {
+              budgetVariability
+              monthlyAmounts {
+                ...BudgetDataMonthlyAmountsFields
+                __typename
+              }
+              __typename
+            }
+            
+            fragment BudgetDataTotalsByMonthFields on BudgetTotals {
+              actualAmount
+              plannedAmount
+              previousMonthRolloverAmount
+              remainingAmount
+              __typename
+            }
+            
+            fragment BudgetTotalsByMonthFields on BudgetMonthTotals {
+              month
+              totalIncome {
+                ...BudgetDataTotalsByMonthFields
+                __typename
+              }
+              totalExpenses {
+                ...BudgetDataTotalsByMonthFields
+                __typename
+              }
+              totalFixedExpenses {
+                ...BudgetDataTotalsByMonthFields
+                __typename
+              }
+              totalNonMonthlyExpenses {
+                ...BudgetDataTotalsByMonthFields
+                __typename
+              }
+              totalFlexibleExpenses {
+                ...BudgetDataTotalsByMonthFields
+                __typename
+              }
+              __typename
+            }
+            
+            fragment BudgetRolloverPeriodFields on BudgetRolloverPeriod {
+              id
+              startMonth
+              endMonth
+              startingBalance
+              targetAmount
+              frequency
+              type
+              __typename
+            }
+            
+            fragment BudgetCategoryFields on Category {
+              id
+              name
+              icon
+              order
+              budgetVariability
+              excludeFromBudget
+              isSystemCategory
+              updatedAt
+              group {
+                id
+                type
+                budgetVariability
+                groupLevelBudgetingEnabled
+                __typename
+              }
+              rolloverPeriod {
+                ...BudgetRolloverPeriodFields
+                __typename
+              }
+              __typename
+            }
+            
+            fragment BudgetDataFields on BudgetData {
+              monthlyAmountsByCategory {
+                ...BudgetMonthlyAmountsByCategoryFields
+                __typename
+              }
+              monthlyAmountsByCategoryGroup {
+                ...BudgetMonthlyAmountsByCategoryGroupFields
+                __typename
+              }
+              monthlyAmountsForFlexExpense {
+                ...BudgetMonthlyAmountsForFlexExpenseFields
+                __typename
+              }
+              totalsByMonth {
+                ...BudgetTotalsByMonthFields
+                __typename
+              }
+              __typename
+            }
+            
+            fragment BudgetCategoryGroupFields on CategoryGroup {
+              id
+              name
+              order
+              type
+              budgetVariability
+              updatedAt
+              groupLevelBudgetingEnabled
+              categories {
+                ...BudgetCategoryFields
+                __typename
+              }
+              rolloverPeriod {
+                id
+                type
+                startMonth
+                endMonth
+                startingBalance
+                frequency
+                targetAmount
+                __typename
+              }
+              __typename
+            }
+            
+            fragment BudgetDataGoalsV2Fields on GoalV2 {
+              id
+              name
+              archivedAt
+              completedAt
+              priority
+              imageStorageProvider
+              imageStorageProviderId
+              plannedContributions(startMonth: $startDate, endMonth: $endDate) {
+                id
+                month
+                amount
+                __typename
+              }
+              monthlyContributionSummaries(startMonth: $startDate, endMonth: $endDate) {
+                month
+                sum
+                __typename
+              }
+              __typename
+            }            
+            """
+)
+
+GET_SUBSCRIPTION_DETAILS = gql(
+    """
+          query GetSubscriptionDetails {
+            subscription {
+              id
+              paymentSource
+              referralCode
+              isOnFreeTrial
+              hasPremiumEntitlement
+              __typename
+            }
+          }
+        """
+)
+
+GET_TRANSACTIONS_PAGE = gql(
+    """
+            query GetTransactionsPage($filters: TransactionFilterInput) {
+              aggregates(filters: $filters) {
+                summary {
+                  ...TransactionsSummaryFields
+                  __typename
+                }
+                __typename
+              }
+            }
+
+            fragment TransactionsSummaryFields on TransactionsSummary {
+              avg
+              count
+              max
+              maxExpense
+              sum
+              sumIncome
+              sumExpense
+              first
+              last
+              __typename
+            }
+        """
+)
+
+GET_TRANSACTIONS_LIST = gql(
+    """
+          query GetTransactionsList($offset: Int, $limit: Int, $filters: TransactionFilterInput, $orderBy: TransactionOrdering) {
+            allTransactions(filters: $filters) {
+              totalCount
+              results(offset: $offset, limit: $limit, orderBy: $orderBy) {
+                id
+                ...TransactionOverviewFields
+                __typename
+              }
+              __typename
+            }
+            transactionRules {
+              id
+              __typename
+            }
+          }
+    
+          fragment TransactionOverviewFields on Transaction {
+            id
+            amount
+            pending
+            date
+            hideFromReports
+            plaidName
+            notes
+            isRecurring
+            reviewStatus
+            needsReview
+            attachments {
+              id
+              extension
+              filename
+              originalAssetUrl
+              publicId
+              sizeBytes
+              __typename
+            }
+            isSplitTransaction
+            createdAt
+            updatedAt
+            category {
+              id
+              name
+              __typename
+            }
+            merchant {
+              name
+              id
+              transactionsCount
+              __typename
+            }
+            account {
+              id
+              displayName
+              __typename
+            }
+            tags {
+              id
+              name
+              color
+              order
+              __typename
+            }
+            __typename
+          }
+        """
+)
+
+COMMON__CREATE_TRANSACTION_MUTATION = gql(
+    """
+          mutation Common_CreateTransactionMutation($input: CreateTransactionMutationInput!) {
+            createTransaction(input: $input) {
+              errors {
+                ...PayloadErrorFields
+                __typename
+              }
+              transaction {
+                id
+              }
+              __typename
+            }
+          }
+
+          fragment PayloadErrorFields on PayloadError {
+            fieldErrors {
+              field
+              messages
+              __typename
+            }
+            message
+            code
+            __typename
+          }
+        """
+)
+
+COMMON__DELETE_TRANSACTION_MUTATION = gql(
+    """
+          mutation Common_DeleteTransactionMutation($input: DeleteTransactionMutationInput!) {
+            deleteTransaction(input: $input) {
+              deleted
+              errors {
+                ...PayloadErrorFields
+                __typename
+              }
+              __typename
+            }
+          }
+  
+          fragment PayloadErrorFields on PayloadError {
+            fieldErrors {
+              field
+              messages
+              __typename
+            }
+            message
+            code
+            __typename
+          }
+        """
+)
+
+GET_CATEGORIES = gql(
+    """
+          query GetCategories {
+            categories {
+              ...CategoryFields
+              __typename
+            }
+          }
+
+          fragment CategoryFields on Category {
+            id
+            order
+            name
+            systemCategory
+            isSystemCategory
+            isDisabled
+            updatedAt
+            createdAt
+            group {
+              id
+              name
+              type
+              __typename
+            }
+            __typename
+          }
+        """
+)
+
+WEB__DELETE_CATEGORY = gql(
+    """
+          mutation Web_DeleteCategory($id: UUID!, $moveToCategoryId: UUID) {
+            deleteCategory(id: $id, moveToCategoryId: $moveToCategoryId) {
+              errors {
+                ...PayloadErrorFields
+                __typename
+              }
+              deleted
+              __typename
+            }
+          }
+
+          fragment PayloadErrorFields on PayloadError {
+            fieldErrors {
+              field
+              messages
+              __typename
+            }
+            message
+            code
+            __typename
+          }
+        """
+)
+
+MANAGE_GET_CATEGORY_GROUPS = gql(
+    """
+          query ManageGetCategoryGroups {
+              categoryGroups {
+                  id
+                  name
+                  order
+                  type
+                  updatedAt
+                  createdAt
+                  __typename
+              }
+          }
+        """
+)
+
+GET_ALL_MERCHANTS = gql(
+    """
+          query GetAllMerchants {
+            byMerchant: aggregates(groupBy: ["merchant"]) {
+              groupBy {
+                merchant {
+                  id
+                  name
+                  __typename
+                }
+                __typename
+              }
+              __typename
+            }
+          }
+        """
+)
+
+WEB__CREATE_CATEGORY = gql(
+    """
+            mutation Web_CreateCategory($input: CreateCategoryInput!) {
+                createCategory(input: $input) {
+                    errors {
+                        ...PayloadErrorFields
+                        __typename
+                    }
+                    category {
+                        id
+                        ...CategoryFormFields
+                        __typename
+                    }
+                    __typename
+                }
+            }
+            fragment PayloadErrorFields on PayloadError {
+                fieldErrors {
+                    field
+                    messages
+                    __typename
+                }
+                message
+                code
+                __typename
+            }
+            fragment CategoryFormFields on Category {
+                id
+                order
+                name
+                systemCategory
+                systemCategoryDisplayName
+                budgetVariability
+                isSystemCategory
+                isDisabled
+                group {
+                    id
+                    type
+                    groupLevelBudgetingEnabled
+                    __typename
+                }
+                rolloverPeriod {
+                    id
+                    startMonth
+                    startingBalance
+                    __typename
+                }
+                __typename
+            }
+            """
+)
+
+GET_HOUSEHOLD_TRANSACTION_TAGS = gql(
+    """
+          query GetHouseholdTransactionTags($search: String, $limit: Int, $bulkParams: BulkTransactionDataParams) {
+            householdTransactionTags(
+              search: $search
+              limit: $limit
+              bulkParams: $bulkParams
+            ) {
+              id
+              name
+              color
+              order
+              transactionCount
+              __typename
+            }
+          }
+        """
+)
+
+WEB__SET_TRANSACTION_TAGS = gql(
+    """
+          mutation Web_SetTransactionTags($input: SetTransactionTagsInput!) {
+            setTransactionTags(input: $input) {
+              errors {
+                ...PayloadErrorFields
+                __typename
+              }
+              transaction {
+                id
+                tags {
+                  id
+                  __typename
+                }
+                __typename
+              }
+              __typename
+            }
+          }
+
+          fragment PayloadErrorFields on PayloadError {
+            fieldErrors {
+              field
+              messages
+              __typename
+            }
+            message
+            code
+            __typename
+          }
+          """
+)
+
+GET_TRANSACTION_DRAWER = gql(
+    """
+          query GetTransactionDrawer($id: UUID!, $redirectPosted: Boolean) {
+            getTransaction(id: $id, redirectPosted: $redirectPosted) {
+              id
+              amount
+              pending
+              isRecurring
+              date
+              originalDate
+              hideFromReports
+              needsReview
+              reviewedAt
+              reviewedByUser {
+                id
+                name
+                __typename
+              }
+              plaidName
+              notes
+              hasSplitTransactions
+              isSplitTransaction
+              isManual
+              splitTransactions {
+                id
+                ...TransactionDrawerSplitMessageFields
+                __typename
+              }
+              originalTransaction {
+                id
+                ...OriginalTransactionFields
+                __typename
+              }
+              attachments {
+                id
+                publicId
+                extension
+                sizeBytes
+                filename
+                originalAssetUrl
+                __typename
+              }
+              account {
+                id
+                ...TransactionDrawerAccountSectionFields
+                __typename
+              }
+              category {
+                id
+                __typename
+              }
+              goal {
+                id
+                __typename
+              }
+              merchant {
+                id
+                name
+                transactionCount
+                logoUrl
+                recurringTransactionStream {
+                  id
+                  __typename
+                }
+                __typename
+              }
+              tags {
+                id
+                name
+                color
+                order
+                __typename
+              }
+              needsReviewByUser {
+                id
+                __typename
+              }
+              __typename
+            }
+            myHousehold {
+              users {
+                id
+                name
+                __typename
+              }
+              __typename
+            }
+          }
+
+          fragment TransactionDrawerSplitMessageFields on Transaction {
+            id
+            amount
+            merchant {
+              id
+              name
+              __typename
+            }
+            category {
+              id
+              name
+              __typename
+            }
+            __typename
+          }
+
+          fragment OriginalTransactionFields on Transaction {
+            id
+            date
+            amount
+            merchant {
+              id
+              name
+              __typename
+            }
+            __typename
+          }
+
+          fragment TransactionDrawerAccountSectionFields on Account {
+            id
+            displayName
+            logoUrl
+            id
+            mask
+            subtype {
+              display
+              __typename
+            }
+            __typename
+          }
+        """
+)
+
+TRANSACTION_SPLIT_QUERY = gql(
+    """
+          query TransactionSplitQuery($id: UUID!) {
+            getTransaction(id: $id) {
+              id
+              amount
+              category {
+                id
+                name
+                __typename
+              }
+              merchant {
+                id
+                name
+                __typename
+              }
+              splitTransactions {
+                id
+                merchant {
+                  id
+                  name
+                  __typename
+                }
+                category {
+                  id
+                  name
+                  __typename
+                }
+                amount
+                notes
+                __typename
+              }
+              __typename
+            }
+          }
+        """
+)
+
+COMMON__SPLIT_TRANSACTION_MUTATION = gql(
+    """
+          mutation Common_SplitTransactionMutation($input: UpdateTransactionSplitMutationInput!) {
+            updateTransactionSplit(input: $input) {
+              errors {
+                ...PayloadErrorFields
+                __typename
+              }
+              transaction {
+                id
+                hasSplitTransactions
+                splitTransactions {
+                  id
+                  merchant {
+                    id
+                    name
+                    __typename
+                  }
+                  category {
+                    id
+                    name
+                    __typename
+                  }
+                  amount
+                  notes
+                  __typename
+                }
+                __typename
+              }
+              __typename
+            }
+          }
+
+          fragment PayloadErrorFields on PayloadError {
+            fieldErrors {
+              field
+              messages
+              __typename
+            }
+            message
+            code
+            __typename
+          }
+        """
+)
+
+WEB__GET_CASH_FLOW_PAGE = gql(
+    """
+          query Web_GetCashFlowPage($filters: TransactionFilterInput) {
+            byCategory: aggregates(filters: $filters, groupBy: ["category"]) {
+              groupBy {
+                category {
+                  id
+                  name
+                  group {
+                    id
+                    type
+                    __typename
+                  }
+                  __typename
+                }
+                __typename
+              }
+              summary {
+                sum
+                __typename
+              }
+              __typename
+            }
+            byCategoryGroup: aggregates(filters: $filters, groupBy: ["categoryGroup"]) {
+              groupBy {
+                categoryGroup {
+                  id
+                  name
+                  type
+                  __typename
+                }
+                __typename
+              }
+              summary {
+                sum
+                __typename
+              }
+              __typename
+            }
+            byMerchant: aggregates(filters: $filters, groupBy: ["merchant"]) {
+              groupBy {
+                merchant {
+                  id
+                  name
+                  logoUrl
+                  __typename
+                }
+                __typename
+              }
+              summary {
+                sumIncome
+                sumExpense
+                __typename
+              }
+              __typename
+            }
+            summary: aggregates(filters: $filters, fillEmptyValues: true) {
+              summary {
+                sumIncome
+                sumExpense
+                savings
+                savingsRate
+                __typename
+              }
+              __typename
+            }
+          }
+        """
+)
+
+WEB__GET_CASH_FLOW_PAGE_2 = gql(
+    """
+          query Web_GetCashFlowPage($filters: TransactionFilterInput) {
+            summary: aggregates(filters: $filters, fillEmptyValues: true) {
+              summary {
+                sumIncome
+                sumExpense
+                savings
+                savingsRate
+                __typename
+              }
+              __typename
+            }
+          }
+        """
+)
+
+WEB__TRANSACTION_DRAWER_UPDATE_TRANSACTION = gql(
+    """
+        mutation Web_TransactionDrawerUpdateTransaction($input: UpdateTransactionMutationInput!) {
+            updateTransaction(input: $input) {
+            transaction {
+                id
+                amount
+                pending
+                date
+                hideFromReports
+                needsReview
+                reviewedAt
+                reviewedByUser {
+                id
+                name
+                __typename
+                }
+                plaidName
+                notes
+                isRecurring
+                category {
+                id
+                __typename
+                }
+                goal {
+                id
+                __typename
+                }
+                merchant {
+                id
+                name
+                __typename
+                }
+                __typename
+            }
+            errors {
+                ...PayloadErrorFields
+                __typename
+            }
+            __typename
+            }
+        }
+
+        fragment PayloadErrorFields on PayloadError {
+            fieldErrors {
+            field
+            messages
+            __typename
+            }
+            message
+            code
+            __typename
+        }
+        """
+)
+
+COMMON__UPDATE_BUDGET_ITEM = gql(
+    """
+          mutation Common_UpdateBudgetItem($input: UpdateOrCreateBudgetItemMutationInput!) {
+            updateOrCreateBudgetItem(input: $input) {
+              budgetItem {
+                id
+                budgetAmount
+                __typename
+              }
+              __typename
+            }
+          }
+        """
+)
+
+WEB__GET_UPCOMING_RECURRING_TRANSACTION_ITEMS = gql(
+    """
+            query Web_GetUpcomingRecurringTransactionItems($startDate: Date!, $endDate: Date!, $filters: RecurringTransactionFilter) {
+              recurringTransactionItems(
+                startDate: $startDate
+                endDate: $endDate
+                filters: $filters
+              ) {
+                stream {
+                  id
+                  frequency
+                  amount
+                  isApproximate
+                  merchant {
+                    id
+                    name
+                    logoUrl
+                    __typename
+                  }
+                  __typename
+                }
+                date
+                isPast
+                transactionId
+                amount
+                amountDiff
+                category {
+                  id
+                  name
+                  __typename
+                }
+                account {
+                  id
+                  displayName
+                  logoUrl
+                  __typename
+                }
+                __typename
+              }
+            }
+        """
+)
+
+LOGIN_MUTATION = gql(
+    """
+            mutation LoginMutation(
+                $email: String!,
+                $password: String!,
+                $totpToken: String,
+                $rememberMe: Boolean
+            ) {
+                login(
+                    email: $email,
+                    password: $password,
+                    totpToken: $totpToken,
+                    rememberMe: $rememberMe
+                ) {
+                    token
+                    user {
+                        id
+                        email
+                        __typename
+                    }
+                    errors {
+                        field
+                        messages
+                        __typename
+                    }
+                    __typename
+                }
+            }
+        """
+)
+
